Create ImmigrationPoliciesData
The USCIS Policy Manual is the agency’s centralized online repository for USCIS’ immigration policies. The Policy Manual is replacing the Adjudicator’s Field Manual (AFM), the USCIS Immigration Policy Memoranda site, and other USCIS policy repositories. The Policy Manual contains separate volumes pertaining to different areas of immigration benefits administered by the agency, such as citizenship and naturalization, adjustment of status, and nonimmigrants. The content is organized into different volumes, parts, and chapters.

The Policy Manual provides transparency of immigration policies and furthers consistency, quality, and efficiency consistent with the USCIS mission. The Policy Manual provides all the latest policy updates; an expanded table of contents; keyword search function; and links to the Immigration and Nationality Act and Code of Federal Regulations, as well as public use forms. The Policy Manual contains tables and charts to facilitate understanding of complex topics. The Policy Manual also contains all historical policy updates.

The Policy Manual contains the official policies of USCIS and assists immigration officers in rendering decisions. The Policy Manual is to be followed by all USCIS officers in the performance of their duties but it does not remove their discretion in making adjudicatory decisions. The Policy Manual does not create any substantive or procedural right or benefit that is legally enforceable by any party against the United States or its agencies or officers or any other person.